+++ src/components/App/UserButton.js
@@ -18,54 +18,6 @@ const StyledSignIn = styled.div`
   }
 `;
 
-const UserCard = styled(MenuCard)`
-  width: 300px;
-  height: 100px;
-  top: 65px;
-  right: 60px;
-  display: flex;
-  flex-direction: column;
-  background-color: #f1fffa; // Card color
-  border: 2px solid #93b7be;
-`;
-
-const UserDetails = styled.div`
-  display: flex;
-  align-self: center;
-  height: 70%;
-  width: 90%;
-  border-bottom: 2px solid #93b7be;
-  & svg {
-    font-size: 4em;
-  }
-`;
-const Names = styled.div`
-  padding: 6px;
-  width: 100%;
-`;
-
-const DisplayName = styled.div`
-  font-size: 1.25em;
-  font-weight: bold;
-`;
-
-const LowerCard = styled.div`
-  display: flex;
-  justify-content: space-around;
-  width: 100%;
-  padding: 3px 0 0 0;
-`;
-
-const SignInOut = styled.div`
-  cursor: pointer;
-`;
-
-const is_user_logged_in = () => {
-  // do we have name in storage?
-  // Yes => logged In
-  // No => Show login
-};
-
 const UserButton = () => {
   function signIn(e) {
     e.preventDefault();
@@ -77,41 +29,15 @@ const UserButton = () => {
     auth.signOut();
     localStorage.clear();
   }
-
-  const username = localStorage.getItem('username');
   const displayName = localStorage.getItem('displayName');
-
-  const [card, setCard] = useState(false);
-
-  const showCard = () => {
-    setCard(!card);
-  };
-
   return (
     <StyledSignIn>
       <VscAccount
-        onClick={showCard}
+        onClick={signIn}
         // onMouseEnter={showCard}
         // onMouseLeave={showCard}
       />
       {displayName}
-      {card ? (
-        <UserCard>
-          <UserDetails>
-            <VscSmiley />
-            <Names>
-              <DisplayName>{displayName}</DisplayName>
-              <div>{username}</div>
-            </Names>
-          </UserDetails>
-          <LowerCard>
-            <div className="status">Status</div>
-            <SignInOut className="signInOut" onClick={signIn}>
-              Sign In
-            </SignInOut>
-          </LowerCard>
-        </UserCard>
-      ) : null}
     </StyledSignIn>
   );
 };
